// src/types/financial.ts
export interface ProfitabilityRatios {
  grossMargin: number;
  operatingMargin: number;
  ebitdaMargin: number;
  netMargin: number;
  roa: number;
  roe: number;
  roce: number;
  roic: number;
}